move out common python array stuff
